feat: refactor theme
