--- libs/shared/theme/src/lib/shadow.ts
@@ -1,7 +1,0 @@
-const shadow = {
-  DEFAULT: '0px 4px 5px #00000033',
-  text: '0px 2px 0px #00000026',
-  'text-small': '0px 1px 0px #00000026',
-};
-
-export default shadow;
